--- a/src/Euclidean_Quiet.cpp
+++ b/src/Euclidean_Quiet.cpp
@@ -228,6 +228,13 @@ extern "C" {
 #define BEAT_POSITION_MAX 15
 #define BEAT_POSITION_DEFAULT 0
 
+/// Index into color palette representing not lighting up an LED
+#define PALETTE_OFF 0;
+/// Index into color palette representing lighting up an LED
+#define PALETTE_ON 1;
+/// Index into color palette representing a blinking LED
+#define PALETTE_BLINK 2;
+
 /// Row of LED Display marked "CH SEL" on panel
 #define LED_CH_SEL_Y 6
 /// Column of LED Display marked "TRIG" on panel
@@ -259,6 +266,15 @@ Encoder Enc3(PIN_ENC_3B, PIN_ENC_3A); // Offset  / O
 // (from LedControl.h library)
 // 1 is maximum number of devices that can be controlled
 LedControl lc = LedControl(PIN_OUT_LED_DATA, PIN_OUT_LED_CLOCK, PIN_OUT_LED_SELECT, 1);
+
+/// Each pixel in the framebuffer indexes into this palette with a 2-bit number, 
+/// which corresponds to one of the `PALLETTE_` constants.
+bool palette[4];
+
+/// Buffer that can be drawn into and manipulated before being drawn to the
+/// hardware display. 2 bits per pixel, so it supports 4 colors. Each color is
+/// an index into `palette`.
+uint16_t framebuffer[LED_ROWS];
 
 /// References one of the three channels
 typedef enum Channel {