--- a/src/Euclidean_Quiet.cpp
+++ b/src/Euclidean_Quiet.cpp
@@ -431,6 +431,10 @@ static inline int eeprom_addr_density(Channel channel);
 static inline int eeprom_addr_offset(Channel channel);
 static void active_channel_set(Channel channel);
 static uint8_t output_channel_led_x(OutputChannel channel);
+#define framebuffer_pixel_on(x, y) (framebuffer_pixel_set(x, y, PALETTE_ON))
+#define framebuffer_pixel_off(x, y) (framebuffer_pixel_set(x, y, PALETTE_OFF))
+#define framebuffer_pixel_blink(x, y) (framebuffer_pixel_set(x, y, PALETTE_BLINK))
+static void framebuffer_pixel_set(uint8_t x, uint8_t y, PaletteColor color);
 static void framebuffer_draw_to_display();
 #define led_pixel_on(x, y) (led_pixel_set(x, y, true))
 #define led_pixel_off(x, y) (led_pixel_set(x, y, false))
@@ -1212,6 +1216,18 @@ static uint8_t output_channel_led_x(OutputChannel channel) {
     result = LED_OUT_OFFBEAT_X;
   }
   return result;
+}
+
+static void framebuffer_pixel_set(uint8_t x, uint8_t y, PaletteColor color) {
+  // Clear existing color
+  uint16_t mask = 0x0003; // Must be 16 bits because it gets inverted
+  framebuffer[y] &= ~(mask << (x * 2));
+
+  // Set new color
+  framebuffer[y] |= (color << (x * 2));
+
+  // Mark as needing redraw
+  framebuffer_row_needs_redraw |= (0x01 << y);
 }
 
 static void framebuffer_draw_to_display() {