… src/config.h
… src/hardware.h
… src/types.h