Instead of drawing directly to the LED matrix display, UI code now draws to a 2-bit (4 color) framebuffer that is then copied to the LED display one row at a time. This has a number of benefits, including:

- Drawing indirectly allows for optimizations around when to set pixels on the LED matrix. Copying 1 row per main loop cycle from the framebuffer to the LED alleviates a bottleneck there. This has brought output latency down significantly (See performance measurements at the end of this description).
- The use of 2 additional colors in the palette beyond simply "off" and "on" enables straightforward implementation of pixel-level visual effects including global blinking and marching ants.
- Indirect drawing also allows for future visual effects such as scrolling, which is planned to be used for an upcoming menu system.

## Performance Measurements

The Euclidean was set up with all three channels having 16 length, Ch 1 & 2 having 9 density, and Ch 3 having 16 density. No parameters were adjusted during tests. All measurements captured with an oscilloscope.  Measurements in the table are worst-observed Trig In -> Channel 3 Output latency, when the module is clocked by various sources to test real-world and extreme settings. 

| Clock Source | Original Firmware  | Now |
| ------------- | ------------- | ------------- |
| **Single Pulse** | 9.1ms | 0.4ms |
| **Korg SQ-1 Sequencer at Fastest Tempo** | 9.1ms | 0.6ms |
| **Audio-Rate Clock (Maximum Supported)** |  80hz (12ms period) | 330hz (3ms period) |

At this point, the maximum audio-rate frequency supported is limited by the length of the output pulses produced.